backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
